Be much smarter about when to render the scene.
This uses the orbit controls start, end, and change events to only draw when we need to, rather than rendering all the time.

This likely will make random scene changes not available until the next update.  By default, we should hook up to the update events to redraw the scene once more.