Work on a bug dealing with getting the appropriate mouse coordinates.
I noticed that that.options.dom.width was much higher than the pixel width of the canvas element.  It appears that it is the difference in the CSS width and the canvas width from doing something like <canvas width="500" style="width:100px">.  The mouse click appears to follow the css width (100px), but if you do canvas.width, you get 500.  So this change calculates the width based on the mouse click coordinate and the CSS width and height (rather than the canvas coordinate width and height).

There still seems to be something wrong with some calculation somewhere---I still can't hit an object by clicking on it.  It may be related to a messed up mouseY calculation.  I'm not sure.