Add a custom helper class CloneArray
Adds a helper class that lets you add clones of a given Object3D. The
location of the clones are determined by the passed array of
`positions`. There are two modes: merge = true or merge = false. When
merge = true, the objects are all merged into one big geometry, reducing
the number of draw calls. The cost is higher memory usage, as the
attribute arrays are replicated equal to the number of clones.